gui: fix an error which caused the buffers can never be released
As 'mMaxAcquiredBuffers' is a 'int32_t' value but 'mCurrentMaxAcquiredBufferCount'
is a 'uint32_t' value, so there is a situation where 'mCurrentMaxAcquiredBufferCount'
is bigger than  'mMaxAcquiredBuffers',  this woudl let 'numPendingBuffersToHold'
(a unsigned value) to becomes a large positive number rather than a negative number.
And that caused the code in the subsequent while loop never be called.

Change-Id: I7f57c52f51efc4760e66c897b4d74923da679d26
Signed-off-by: chengdaozhong <chengdaozh@gmail.com>
Signed-off-by: Jason Edson <jaysonedson@gmail.com>